de_technical_test_batch/requirements.txt
faker
pandas
sqlalchemy
apache-airflow-providers-postgres
pendulum